ensure multiplication takes place in unsigned long
the cast here seems dobious. the multiplication will take place in
`unsigned int` and the cast only affects the _result_ and not the
multiplication itself. which means that the high 32bits will remain zero
and the shift will thus always produce zero as well.

use UL suffix on the integer constant to ensure that the multiplication
itself happens at `unsigned long`.

+++ xbanish.c
@@ -569,8 +569,8 @@ set_alarm(XSyncAlarm *alarm, XSyncTestType test)
 	attr.trigger.counter = idler_counter;
 	attr.trigger.test_type = test;
 	attr.trigger.value_type = XSyncRelative;
-	XSyncIntsToValue(&attr.trigger.wait_value, timeout * 1000,
-	    (unsigned long)(timeout * 1000) >> 32);
+	XSyncIntsToValue(&attr.trigger.wait_value, timeout * 1000UL,
+	    (timeout * 1000UL) >> 32);
 	XSyncIntToValue(&attr.delta, 0);
 
 	flags = XSyncCACounter | XSyncCATestType | XSyncCAValueType |
